+++ src/main/java/com/amine/pfe/georef_module/gcp/service/port/GcpFactory.java
@@ -4,5 +4,5 @@ import com.amine.pfe.georef_module.entity.Gcp;
 import com.amine.pfe.georef_module.entity.GeorefImage;
 
 public interface GcpFactory {
-    Gcp createGcp(GeorefImage image, int sourceX, int sourceY, double mapX, double mapY, Integer index);
+    Gcp createGcp(GeorefImage image, Double sourceX, Double sourceY, double mapX, double mapY, Integer index);
 }
